requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic>=2.8.2,<3
pydantic-settings>=2.3.1,<3
SQLAlchemy==2.0.31
alembic==1.13.2
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.9
httpx==0.27.0
azure-storage-blob==12.21.0
structlog==24.1.0
prometheus-fastapi-instrumentator==6.1.0
slowapi==0.1.9
python-slugify==8.0.4

